remove dataprocessorinput port
