+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/neo/WorkerInternalQueue.scala
@@ -7,6 +7,7 @@ import edu.uci.ics.amber.engine.architecture.worker.neo.WorkerInternalQueue.{
   InternalQueueElement,
   SenderTuplePair
 }
+import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.ambertag.LayerTag
 import edu.uci.ics.amber.engine.common.tuple.ITuple
 
@@ -44,6 +45,10 @@ class WorkerInternalQueue {
   // TODO: we also need to refactor all identifiers
   var inputMap = new mutable.HashMap[LayerTag, Int]
 
+  // indicate if all upstreams exhausted
+  private var allExhausted = false
+  private var inputExhaustedCount = 0
+
   /** take one FIFO batch from worker actor then put into the queue.
     * @param batch
     */
@@ -70,4 +75,31 @@ class WorkerInternalQueue {
   def addDummyInput(): Unit = {
     blockingDeque.add(DummyInput())
   }
+
+  /** get next input tuple
+    * should only be called from dp thread
+    * @return tuple
+    */
+  def getNextInputPair: (Int, Either[ITuple, InputExhausted]) = {
+
+    val currentInput = blockingDeque.take()
+
+    currentInput match {
+      case SenderTuplePair(senderRef, tuple) =>
+        // if current batch is a data batch, return tuple
+        // empty iterators will be filtered in WorkerInternalQueue so we can safely call next()
+        (senderRef, Left(tuple))
+      case EndMarker(senderRef) =>
+        // current batch is an End of Data sign.
+        inputExhaustedCount += 1
+        // check if End of Data sign from every upstream has been received
+        allExhausted = inputMap.size == inputExhaustedCount
+        (senderRef, Right(InputExhausted()))
+      case DummyInput() =>
+        // if the batch is dummy batch inserted by worker, return null to unblock dp thread
+        (-1, null)
+    }
+  }
+
+  def isAllUpstreamsExhausted: Boolean = allExhausted
 }
